+++ Day/05/src/console/Domain/Almanac.cs
@@ -2,5 +2,5 @@ using System.Collections.Immutable;
 
 namespace Domain;
 
-record Almanac(ImmutableList<int> SeedList,
+record Almanac(ImmutableList<long> SeedList,
                ImmutableDictionary<MapKind, Map> Maps);
